implementa entidades con sus anotaciones.
