--- src/main/java/com/felipelopez/Biblioteca/entity/Autor.java
@@ -1,8 +1,0 @@
-package com.felipelopez.Biblioteca.entity;
-
-public class Autor {
-
-    public Long idAutor;
-
-    public String nombreAutor;
-}
