--- src/main/java/com/felipelopez/Biblioteca/entity/FechaPublicacion.java
@@ -1,8 +1,0 @@
-package com.felipelopez.Biblioteca.entity;
-
-public class FechaPublicacion {
-
-        public Long idFechaPublicacion;
-
-        public String fechaPublicacion;
-}
