--- src/main/java/com/felipelopez/Biblioteca/entity/Libros.java
@@ -1,8 +1,0 @@
-package com.felipelopez.Biblioteca.entity;
-
-public class Libros {
-
-    public Long idLibro;
-
-    public String tituloLibro;
-}
