+++ src/main/java/com/felipelopez/Biblioteca/model/entity/Autor.java
@@ -1,0 +1,21 @@
+package com.felipelopez.Biblioteca.model.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "autor")
+public class Autor {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    public Long idAutor;
+
+    @Column(name = "nombre_autor", nullable = false)
+    public String nombreAutor;
+}
