+++ src/main/java/com/felipelopez/Biblioteca/model/entity/Libro.java
@@ -1,0 +1,21 @@
+package com.felipelopez.Biblioteca.model.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "libro")
+public class Libro {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    public Long idLibro;
+
+    @Column(name = "titulo_libro", nullable = false)
+    public String tituloLibro;
+}
